テスト用広告を本番用に変えた 課金テストも終わった アイコンの色を変えたくなった これ終わったら完成だ。がんばろう

+++ BodyImpersonator/tappedImageVC.m
@@ -50,7 +50,7 @@
     countViewChanged ++;
     [[NSUserDefaults standardUserDefaults] setInteger:countViewChanged forKey:@"KEY_countUpViewChanged"];
     [[NSUserDefaults standardUserDefaults]synchronize];
-    NSLog(@"viewchanged:%d",countViewChanged);
+    NSLog(@"viewchanged:%ld",(long)countViewChanged);
     
       _kIndicator = [kBIIndicator alloc];
 }
